requirements.txt
Flask==1.0.2
psycopg2-binary==2.8.4
flask_restplus==0.11.0
webargs==6.0.0
Werkzeug==0.14.1
uwsgi==2.0.18
marshmallow==3.2.1
SQLAlchemy==1.2.19
dramatiq[watch]==1.8.1
pika==1.1.0
apispec==3.3.0
pytest>=5.2.2,<6
pytest-mock>=1.11.2,<2
factory-boy==2.12.0